+++ .eslintignore
@@ -5,6 +5,5 @@ www/
 icons/
 dependent-usage-analyzer/
 build-scss.js
-build-css.js
 component-generator/
 example/
