+++ build-scss.js
@@ -1,7 +1,6 @@
-var path = require('path');
-var sass = require('sass');
-var fs = require('fs');
-
+const path = require('path');
+const sass = require('sass');
+const fs = require('fs');
 
 // Resolve tildas the way webpack does
 var tildaImporter = function(url, prev, done) {
@@ -12,14 +11,35 @@ var tildaImporter = function(url, prev, done) {
   return { file: url };
 };
 
+const compileStyleSheets = (path, output = 'expanded') => {
+  return sass.renderSync({
+    file: path,
+    outputStyle: output,
+    importer: tildaImporter,
+  });
+};
 
-// Core paragon style
-var coreResult = sass.renderSync({
-  file: './scss/core/core.scss',
-  outputStyle: 'compressed',
-  importer: tildaImporter,
-});
+fs.writeFileSync(
+    './dist/core.css',
+    compileStyleSheets('./scss/core/core.scss').css
+);
 
-fs.writeFileSync('./dist/paragon.css', coreResult.css);
+fs.writeFileSync(
+    './dist/core.min.css',
+    compileStyleSheets('./scss/core/core.scss', 'compressed').css
+);
 
-module.exports = tildaImporter;
+const compileThemeStyleSheets = (themeVariant) => {
+  fs.writeFileSync(
+      `./dist/${themeVariant}.css`,
+      compileStyleSheets(`./scss/core/css/${themeVariant}/variables.css`).css
+  );
+  fs.writeFileSync(
+      `./dist/${themeVariant}.min.css`,
+      compileStyleSheets(`./scss/core/css/${themeVariant}/variables.css`, 'compressed').css
+  );
+};
+
+const THEME_VARIANTS = ['light'];
+THEME_VARIANTS.forEach(themeVariant => compileThemeStyleSheets(themeVariant));
+
