+++ scss/core/css/light/variables.css
@@ -1,7 +1,7 @@
 /**
  * IMPORTANT: This file is the result of assembling design tokens
  * Do not edit directly
- * Generated on Tue, 28 Feb 2023 13:23:54 GMT
+ * Generated on Wed, 01 Mar 2023 17:14:23 GMT
  */
 
 :root {
